Generalize fetching.
